Phase 2 – Problem 6 – Correction
The max_scoring_num_rolls function now returns the number of dice
rolls that results in the highest score on average.
Instead of collecting the results in a single string and then printing it
at the very end, prints are made as the results are available. This solves
the newline problem previously encountered.